Example terraform scripts for installing MKE on GCP
This is an example of terraform scripts that prepares a cluster
on Google Cloud Platform to install Mirantis Kubernetes Engine.

It creates following GCP resources:
 - A VPC with specified MTU (default value is 1500) and one subnetwork
 - Linux and Windows instances using public images provided by GCP
 - Firewall rules to allow internal and external traffic
 - A network load balancers for MKE and MSR that targets corresponding  unmanaged instance group
(along with health checks)

MTU:
The default MTU is set to 1500 (rather than the default 1460). This is to an issue where swarm services
fails because Docker network uses 1500 as the default MTU.

Known Issues:
* MSR (Mirantis Secure Registry) installation fails due to invalid SSL certificate for the load balancer IP
* Windows nodes does not work when it is assigned "Kubernetes" orchestration